Refactor/쿼리 최적화를 위한 Review 테이블 구조 변경 (#228)
1. 댓글에 좋아요를 누르는 행위보다 댓글을 조회하는 행위가 일반적으로 더 많이 발생할 것으로 예상됩니다. 이에 따라 조회 부분을 최적화하는게 성능 상 이점을 누릴 수 있지 않을까 고민했습니다.

기존의 구조
댓글 좋아요를 누르면 insert Query 가 한 번 발생합니다.
댓글 조회 시, 좋아요 수를 구해야 하기 때문에 카운트 쿼리가 상품에 달린 댓글 수 만큼 나가게 됩니다.

변경된 구조
리뷰 엔티티에서 좋아요 총 수를 필드로 가지고 있습니다. 이에 따라
댓글에 첫 좋아요를 누를 시 총 4번의 쿼리가 발생합니다. 이후 좋아요/좋아요 취소 요청을 보낼 경우 3번의 쿼리가 발생합니다. 이에 따라 응답 속도가 지연될 것으로 예상됩니다.

다만, 더 이상 좋아요 카운트 쿼리가 발생하지 않게 됩니다.

+++ src/main/java/com/jxx/xuni/review/dto/response/ReviewOneResponse.java
@@ -12,6 +12,6 @@ public record ReviewOneResponse(
         Long reviewerId,
         String reviewerName,
         Progress progress,
-        Long likeCnt
+        Integer likeCnt
 ) {
 }
